add support to multi line string
This commit changed how the lexer behaves. Now in the first step of the
lexer `lex_file` we just consumes the lines of the file and send it
directly to `lex_lines`. I introduced a new main loop inside the
`lex_lines` so we can walk line by line. I checked that the error
reporting seems to be working fine.