get rid of `sim` subcommand in favor of `com -r`
This commit removes completely the simulation subcomand `sim` from
Porth. We also introduces a new subcommand `run` which is just an alias
to `com -r -s`.

Now we can use the run command like this:
```
$ cat program.porth
34 35 + print
$ ./porth.py run program.porth
69
```

I fixed the `tests.py` to reflect these changes.

I also introduced a better error report in `cmd_call_echoed` so we can
report a better error message in case the user is missing something to
run the command.