set `basedir` to current working directory if it is empty
This commit fix the bug when `basedir` is an empty string.

To reproduce this bug imagine the following scenario:

```
$ cat test.porth
34 35 + print
$ ./porth.py com -r test.py
$
```

As you can see the expected output should be `69` but it won't print
anything. This happens because `basedir` is empty and `basepath` is
`["test"]` so `exit(cmd_call_echoed([basepath] + argv, silent))` will
run the command "test" which is a valid command in $PATH.