Add switch blocks for simulation and compile modes
Added switch blocks using ints. This also adds the switch, case, and break keywords. In compilation a dictionary is created in the switch's operand. The None key contains a tuple with the location of the default case (-1 if none exists) and the location of the end of the switch block. The other elements of the dictionary are the case conditions. In simulation mode the dictonary is indexed to jump to the matching case. In compilation mode all the conditions are check consecutively until a matching one is found. If no matching one is found there is a non conditional jump to either the default case or the end of the switch block. If a case is found without anything on the stack it will be treated as the default case. Since only one default case is possible having more will cause a compile error. A default keyword could be easily added but I wanted to minimize the amount of new keywords added.